requirements.txt
PyYAML==6.0.1
packaging==23.2
dacite==1.8.1
apscheduler==3.10.4
pytz==2024.1
confluent-kafka==2.3.0
jsonpath-ng==1.6.1
requests==2.31.0
genson==1.2.2
lazy-write==0.0.1
case-converter==1.1.0
jsonschema==4.21.1
kubernetes==29.0.0
avro-python3==1.10.2
ncclient==0.6.15
paramiko==3.4.0
Flask==2.2.5
prometheus-client==0.20.0
./vmware_tcsa_collector_sdk_python-0.0.38-py3-none-any.whl